Trump Anki app's `body` styles
It's a bit icky to have to do this, and it could always be moved to the specific location it's needed (notes etc) ... but put it in `_root.less` for now.

- Move @font-face outside `:root`
- Create a `partials/_root.less` file

This is the same setup as print-first-css now. Copy everything from `:root` to use `body` to override Anki defaults.

Closes #117