Reorder the `--color-card-` variables
The original method in `print-first-css` was to have `l` as the _least_ lightest and `llll` as the _most_ light.

It might make sense to change these variable names in future as `--color-b` isn't actually used and is  more WHITE than black.

Create an issue to look at Material's design tokens which seem more specific.