Add `.nightmode` for the simple template
The code is getting a little bit narly, but we've added in the `.nightmode` or "dark mode" for the simple card.

I'm beginning to think it would be wise to overhaul the card classes completely, ready for a potential Draw! card.

We're basically replicating the exact same card for both cards, with different `.simple-` and `.missing-` prefixes.

A few minor bits removed too.