Change test classes to `#demo` and other things
Now that the `Master` files seem to work properly, and I've tested them with the `Missing!` fields as well, we can get rid of the old card types and replace them with the simplified and unified CSS.

- Rename `/master` to `/simple`
- Ditch the old `/simple` folder
- Ditch the old `/missing` folder
- Add the `.cloze` code
- Add some other necessary CSS
- Check it compiles ok (it does)
- Remove a `testing.html` file

It's almost impossible to share the master template between the two card types, as `#answer` is in two different places depending on the card type.

You could wrap everything with the `{{# ... }} ... {{/ ... }}` tags to make things conditional, but it's likely to get messy fast.

Just stick to isolating the cards, but share the CSS as much as possible.