Rewrite the GPS doc
It's a bit more coherent with questions asked and some answers. Probably need help with this from the GPS people (if they're still available!)

TL;DR in most cases it's fair enough, but there's some nuance, inheritance issues, and fringe cases that I'm not sure what to do with.

What happens, for example, when an element is _almost_ identical across pages?