Update code in the data files
I have a feeling it's best to keep the data files and the demo files in sync, but I often use the data files myself to create my own cards, and it's a bit annoying having to keep on "undoing" the changes I make (to reset to more simpler code defaults).

How can I manage this? Clone it myself?!