Remove chance for human error
At least one person has tried to add a rendered Markdown fenced code block into the `code string` field, where it is not allowed. Make this more explicit.

There's not way to properly enforce it with Anki really — and if I change it to the visual editor for that field, it's likely to add junk html.