Maintain consistency in punctuation
At some point I'm going to have to formalise a styleguide, like the issue in Print-First-CSS:

https://github.com/badlydrawnrob/print-first-css/issues/10

https://service-manual.ons.gov.uk/content/formatting-and-punctuation/lists

As most of the use of bullet points is a list (and not individual statements) remove all full stops and commas.

Keep in the exclamation marks for emphasis, and question marks for questions.

As for CSS files ... I'm not _that_ pedantic! Often they're numbered lists and each list item is an independant sentence.